[REFACT] 호출 방식 변경
